Return unauthorized on failed login to prevent http auth dialog
Chrome shows an http basic auth dialog when receiving an 401
Authentication required reply which contains the "WWW-Authenticate"
header. The previous workaround with explicitly providing an username
and password did not work for localhost.

By returning Unauthorized within Servant the browser gets an 403 reply
for which no basic auth dialog will be shown.