Strip WWW-Authenticate header
From #9:

"Chrome shows an http basic auth dialog when receiving an 401
Authentication required reply which contains the "WWW-Authenticate"
header. The previous workaround with explicitly providing an username
and password did not work for localhost."

This header is of no use to adora-belle itself, so simply strip it
globally using a wai middleware.