Security: Use Data.ByteArray.constEq for password comparision
This fixes a (potential) timing sidechannel in the password checking by
using a constant-time stringequality check.  This only concerns the
plaintext, non-hashed password backend.  Please note that the fix is
(due to the nature of variable-length user passwords) partial in that it
still leaks information on the password length with the plaintext
backend.